+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Add `WSDirector::Snapshot`. ([@palkan][])
+
 - Add `locals` support when running scenarios programmatically. ([@palkan][])
 
 - Add partial receive data matching support (via `data>` field). ([@palkan][])